+++ Assets/Script/GamePlay/Ennemy/TanksEnnemy.cs
@@ -47,7 +47,7 @@ public class TanksEnnemy : EnnemyBehavior
 
     private IEnumerator Reset()
     {
-        yield return new WaitForSeconds(60 / AudioHelmClock.Instance.bpm * 0.5f);
+        yield return new WaitForSeconds(60 / (float)AudioHelmClock.Instance.bpm * 0.5f);
         cptDead =  0 ;
         canReset = false;
     }
